authentication

+++ src/Business/User.business.ts
@@ -1,13 +1,17 @@
+import * as dayjs from "dayjs";
 import { validate, ValidationError } from "class-validator";
 import { getRepository, Repository } from "typeorm";
+import { RefreshToken } from "../models/RefreshToken";
 import { User } from "../models/User";
+import { GenerateRefreshToken } from "../provider/GenerateRefreshToken";
 import { HashManager } from "../services/hashManager";
-import { Authenticator } from "../services/TokenGenerator";
+import { authenticationData, Authenticator } from "../services/TokenGenerator";
 import {
   UserRequest,
   UserResponse,
   UserRequestById,
   UserRequestLogin,
+  UserReturnToken,
 } from "../services/types";
 
 export class UserBusiness {
@@ -46,7 +50,7 @@ export class UserBusiness {
   userLogin = async ({
     email,
     password,
-  }: UserRequestLogin): Promise<UserResponse | Error> => {
+  }: UserRequestLogin): Promise<any | Error> => {
     const repository: Repository<User> = getRepository(User);
     const user: User = await repository.findOne({ email });
     if (!user) return new Error("User not found.");
@@ -62,9 +66,13 @@ export class UserBusiness {
       id: user.id,
     });
 
-    const { name, nickname } = user;
+    const genereteRefreshToken: GenerateRefreshToken =
+      new GenerateRefreshToken();
+    const refreshToken: RefreshToken = await genereteRefreshToken.execute(
+      user.id
+    );
 
-    return { user: { name, nickname, email }, token };
+    return { token, refreshToken };
   };
 
   read = async ({ id }: UserRequestById): Promise<User | Error> => {
@@ -74,15 +82,18 @@ export class UserBusiness {
     return findUserById;
   };
 
-  update = async ({
-    id,
-    name,
-    nickname,
-    email,
-    password,
-  }: UserRequest): Promise<User | Error> => {
+  update = async (
+    token: string,
+    { id, name, nickname, email, password }: UserRequest
+  ): Promise<User | Error> => {
     const repository: Repository<User> = getRepository(User);
     const updateUser: User = await repository.findOne({ id });
+    const accessToken: authenticationData = new Authenticator().getTokenData(
+      token
+    );
+
+    if (accessToken === null) return new Error("invalid token");
+
     if (!updateUser) return new Error("provided id not found.");
     updateUser.name = name ? name : updateUser.name;
     updateUser.nickname = nickname ? nickname : updateUser.nickname;
@@ -92,10 +103,57 @@ export class UserBusiness {
     await repository.save(updateUser);
   };
 
-  delete = async ({ id }: UserRequestById): Promise<Error | User> => {
+  delete = async (
+    token: string,
+    { id }: UserRequestById
+  ): Promise<Error | User> => {
     const repository: Repository<User> = getRepository(User);
+
+    const accessToken: authenticationData = new Authenticator().getTokenData(
+      token
+    );
+
+    if (accessToken === null) return new Error("invalid token");
+
     const deleteUserById: User | Error = await repository.findOne({ id });
     if (!deleteUserById) return new Error("provided id not found.");
     await repository.delete({ id });
+  };
+
+  refreshToken = async (
+    refresh_token: string
+  ): Promise<UserReturnToken | Error | UserResponse> => {
+    const refreshTokenRepository = getRepository(RefreshToken);
+    const findTokenId = await refreshTokenRepository.findOne({
+      id: refresh_token,
+    });
+    if (!findTokenId) return new Error("Refresh token invalid");
+
+    const accessToken = getRepository(RefreshToken);
+    await accessToken.delete({ user_id: findTokenId.id });
+
+    const token: string = new Authenticator().generateToken({
+      id: findTokenId.user_id,
+    });
+
+    const refreshTokenExpired: boolean = dayjs().isAfter(
+      dayjs.unix(findTokenId.expiresIn)
+    );
+
+    const newRepository: Repository<RefreshToken> = getRepository(RefreshToken);
+    if (refreshTokenExpired) {
+      await newRepository.delete({
+        id: findTokenId.user_id,
+      });
+
+      const newRefreshToken: GenerateRefreshToken = new GenerateRefreshToken();
+      const resultRefreshToken: RefreshToken = await newRefreshToken.execute(
+        findTokenId.id
+      );
+
+      return { findTokenId, resultRefreshToken };
+    }
+
+    return { token };
   };
 }
